requirements.txt
amqp==1.4.9
anyjson==0.3.3
appenlight-client==0.6.16
appnope==0.1.0
autopep8==1.2.2
billiard==3.3.0.22
boto==2.39.0
celery==3.1.20
coverage==4.0.3
decorator==4.0.9
dj-database-url==0.4.0
Django==1.9.2
django-debug-toolbar==1.4
django-filter==0.12.0
django-naomi==0.8
django-queryinspect==0.1.0
django-silk==0.5.6
django-storages==1.4
djangorestframework==3.3.2
gnureadline==6.3.3
gunicorn==19.4.5
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
Jinja2==2.8
kombu==3.0.33
line-profiler==1.0
MarkupSafe==0.23
mock==1.3.0
opbeat==3.2.2
path.py==8.1.2
pbr==1.8.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.2
python-dateutil==2.5.0
python-decouple==3.0
pytz==2015.7
pyzmq==15.2.0
requests==2.9.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.18
traitlets==4.1.0
WebOb==1.5.1
whitenoise==2.0.6
yet-another-django-profiler==1.0.3
